requirements.txt
python-telegram-bot==20.7
httpx==0.25.2
python-dotenv==1.0.1
asyncpg==0.29.0
aiohttp==3.9.1
nest-asyncio==1.6.0
requests==2.28.2
redis>=7.1.0
